Struct patricia_merkle::modified_patricia_merkle::ModifiedPatriciaMerkleTrie

pub struct ModifiedPatriciaMerkleTrie<S> { /* fields omitted */ }

Modified Patricia Merkle Trie implementation. S is the schema of the underlying DB that stores the modified patricia merkle trie.

Methods

impl<S> ModifiedPatriciaMerkleTrie<S> where
S: PatriciaMerkleSchema,

pub fn new(db: Arc<DB<S>>) -> Self

Constructs a new ModifiedPatriciaMerkleTrie instance.

pub fn put<T>(
&self,
kvs: Vec<(Nibbler, T)>
) -> Result<(SchemaBatchWithIndex<S>, HashValue)> where
T: Into<Vec<u8>>,
Vec<u8>: From<T>,

Puts a list of key/value pairs into the P-M tree. All of them are supposed to come from a single transaction. Returns a SchemaBatchWithIndex and the new root hash. The caller is supposed to take the return value and write additional info about root node and latest version with a version number into the write batch and send it to DB.

pub fn get_with_proof<T>(
&self,
nibbler: Nibbler,
version: Version
) -> Result<Option<(T, SparseMerkleProofDeprecated)>> where
T: TryFrom<Vec<u8>>,
<T as TryFrom<Vec<u8>>>::Error: Fail,

Retrieves value by traversing the trie with given key with proof

pub fn validate(&self, root_digest: &HashValue) -> Validity

This method will validate the data integrity of the subtree, whose root node is specified by root_digest.

Returns Validity::Valid if the subtree is valid; Validity::Invalid if the subtree is invalid, so user could retrieve the whole subtree again; Validity::ConditionalValid(invalid_digests) if the subtree is valid on the condition that the nodes with invalid_digests are all valid but they failed to be validated locally. Under such scenario, user may want to just re-fetch smaller subtrees originating from them instead of the whole subtree to save network traffic. Note: if the number of invalid_digests exceeds a limit, thought rarely, due to memory allocation concern, we regard the whole subtree is invalid.